INC-54 refactor: extract logout button into a separate component
