+++ src/components/ui/sidebar/logoutButton/index.ts
@@ -1,0 +1,1 @@
+export * from './LogoutButton'
