+++ src/components/ui/sidebar/logoutButton/logoutButton.module.scss
@@ -1,0 +1,8 @@
+.text {
+  margin-bottom: 15px;
+}
+
+.buttonWrapper {
+  display: flex;
+  justify-content: space-between;
+}
