+++ src/components/ui/sidebar/sidebar.module.scss
@@ -7,12 +7,3 @@
 
   border-right: 1px solid var(--color-dark-300);
 }
-
-.text {
-  margin-bottom: 15px;
-}
-
-.buttonWrapper {
-  display: flex;
-  justify-content: space-between;
-}
